# Conflicts:
#	csv_parser/out/EMSI/EMSI.schema.docx
#	csv_parser/out/EMSI/EMSI.uml_diagram.pdf
#	csv_parser/out/GEO-POS/GEO-POS.schema.docx
#	csv_parser/out/GEO-POS/GEO-POS.uml_diagram.pdf
#	csv_parser/out/GEO-REQ/GEO-REQ.schema.docx
#	csv_parser/out/GEO-REQ/GEO-REQ.uml_diagram.pdf
#	csv_parser/out/GEO-RES/GEO-RES.schema.docx
#	csv_parser/out/GEO-RES/GEO-RES.uml_diagram.pdf
#	csv_parser/out/RC-EDA/RC-EDA.schema.docx
#	csv_parser/out/RC-EDA/RC-EDA.uml_diagram.pdf
#	csv_parser/out/RC-REF/RC-REF.schema.docx
#	csv_parser/out/RC-REF/RC-REF.uml_diagram.pdf
#	csv_parser/out/RPIS/RPIS.schema.docx
#	csv_parser/out/RPIS/RPIS.uml_diagram.pdf
#	csv_parser/out/RS-DR/RS-DR.schema.docx
#	csv_parser/out/RS-DR/RS-DR.uml_diagram.pdf
#	csv_parser/out/RS-EDA-MAJ/RS-EDA-MAJ.schema.docx
#	csv_parser/out/RS-EDA-MAJ/RS-EDA-MAJ.uml_diagram.pdf
#	csv_parser/out/RS-EDA/RS-EDA.schema.docx
#	csv_parser/out/RS-EDA/RS-EDA.uml_diagram.pdf
#	csv_parser/out/RS-ERROR/RS-ERROR.schema.docx
#	csv_parser/out/RS-ERROR/RS-ERROR.uml_diagram.pdf
#	csv_parser/out/RS-INFO/RS-INFO.schema.docx
#	csv_parser/out/RS-INFO/RS-INFO.uml_diagram.pdf
#	csv_parser/out/RS-RI/RS-RI.schema.docx
#	csv_parser/out/RS-RI/RS-RI.uml_diagram.pdf
#	csv_parser/out/RS-RR/RS-RR.schema.docx
#	csv_parser/out/RS-RR/RS-RR.uml_diagram.pdf
#	csv_parser/out/RS-SR/RS-SR.schema.docx
#	csv_parser/out/RS-SR/RS-SR.uml_diagram.pdf
#	csv_parser/out/customContent/customContent.schema.docx
#	csv_parser/out/customContent/customContent.uml_diagram.pdf

+++ src/main/java/com/hubsante/model/health/PatientDetail.java
@@ -34,6 +34,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.fasterxml.jackson.dataformat.xml.annotation.*;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Arrays;
 import java.util.Objects;
@@ -49,10 +50,10 @@ import java.util.Objects;
 
 public class PatientDetail {
   public static final String JSON_PROPERTY_WEIGHT = "weight";
-  private Integer weight;
+  private BigDecimal weight;
 
   public static final String JSON_PROPERTY_HEIGHT = "height";
-  private Integer height;
+  private BigDecimal height;
 
   public static final String JSON_PROPERTY_AGE = "age";
   private String age;
@@ -99,7 +100,7 @@ public class PatientDetail {
 
   public PatientDetail() {}
 
-  public PatientDetail weight(Integer weight) {
+  public PatientDetail weight(BigDecimal weight) {
 
     this.weight = weight;
     return this;
@@ -112,17 +113,17 @@ public class PatientDetail {
   @JsonProperty(JSON_PROPERTY_WEIGHT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
 
-  public Integer getWeight() {
+  public BigDecimal getWeight() {
     return weight;
   }
 
   @JsonProperty(JSON_PROPERTY_WEIGHT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setWeight(Integer weight) {
+  public void setWeight(BigDecimal weight) {
     this.weight = weight;
   }
 
-  public PatientDetail height(Integer height) {
+  public PatientDetail height(BigDecimal height) {
 
     this.height = height;
     return this;
@@ -135,13 +136,13 @@ public class PatientDetail {
   @JsonProperty(JSON_PROPERTY_HEIGHT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
 
-  public Integer getHeight() {
+  public BigDecimal getHeight() {
     return height;
   }
 
   @JsonProperty(JSON_PROPERTY_HEIGHT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setHeight(Integer height) {
+  public void setHeight(BigDecimal height) {
     this.height = height;
   }
